lesson-7: custom di moved to di.custom package, created di.dagger package with MyModule.kt
