+++ app/src/main/java/com/rogoz208/mygithubclient/di/custom/DependencyFactory.kt
@@ -1,4 +1,4 @@
-package com.rogoz208.mygithubclient.di
+package com.rogoz208.mygithubclient.di.custom
 
 interface DependencyFactory<T> {
     fun get(): T
