+++ app/src/main/java/com/rogoz208/mygithubclient/di/custom/Factory.kt
@@ -1,4 +1,4 @@
-package com.rogoz208.mygithubclient.di
+package com.rogoz208.mygithubclient.di.custom
 
 class Factory<T>(val create: () -> T) : DependencyFactory<T> {
     override fun get(): T {
