+++ app/src/main/java/com/rogoz208/mygithubclient/di/dagger/MyModule.kt
@@ -1,0 +1,5 @@
+package com.rogoz208.mygithubclient.di.dagger
+
+class MyModule {
+
+}
