+++ app/src/main/java/com/rogoz208/mygithubclient/ui/screens/repositories/RepositoriesActivity.kt
@@ -12,21 +12,24 @@ import by.kirich1409.viewbindingdelegate.viewBinding
 import com.bumptech.glide.Glide
 import com.rogoz208.mygithubclient.R
 import com.rogoz208.mygithubclient.databinding.ActivityRepositoriesBinding
-import com.rogoz208.mygithubclient.di.get
-import com.rogoz208.mygithubclient.di.inject
 import com.rogoz208.mygithubclient.domain.entities.RepositoryEntity
 import com.rogoz208.mygithubclient.domain.repos.RepositoriesRepo
 import com.rogoz208.mygithubclient.ui.screens.repositories.recycler.OnRepositoryClickListener
 import com.rogoz208.mygithubclient.ui.screens.repositories.recycler.RepositoriesAdapter
 import com.rogoz208.mygithubclient.ui.screens.repositories.recycler.RepositoriesDiffCallback
+import javax.inject.Inject
 
 const val USER_EXTRA_KEY = "USER_EXTRA_KEY"
 
 class RepositoriesActivity : AppCompatActivity(R.layout.activity_repositories) {
 
     private val binding by viewBinding(ActivityRepositoriesBinding::bind)
+
+    @Inject
+    private lateinit var repositoriesRepo: RepositoriesRepo
+
     private val viewModel: RepositoriesContract.ViewModel by viewModels {
-        RepositoriesViewModelFactory(get())
+        RepositoriesViewModelFactory(repositoriesRepo)
     }
 
     private val adapter by lazy { RepositoriesAdapter() }
